requirements.txt
# ml/requirements.txt
# Core AI libraries
# deepface
# face_recognition # Alternative/complementary to deepface, choose one or combine as needed
Pillow # For image processing
numpy # For numerical operations, especially with embeddings
scikit-learn # For cosine similarity or other utilities
# faiss-cpu # Or faiss-gpu if a GPU is available, for efficient similarity search
# sentence-transformers # For text embeddings
# torch # For CLIP and sentence-transformers
# transformers # For CLIP and sentence-transformers
# opencv-python # Potentially usfed by deepface/face_recognition
python-multipart # If handling file uploads in FastAPI
# pymongo # For interacting with MongoDB directly from ML scripts if needed, though FastAPI will be the primary interface

# Development/Utility
uvicorn # ASGI server for FastAPI
python-dotenv # For environment variables
passlib[bcrypt] # For password hashing
python-jose[cryptography] # For JWT (JSON Web Tokens)

# Speech-to-Text
#openai-whisper # For local Whisper model

uvicorn
# opencv-python
python-jose[cryptography]
passlib[bcrypt]
python-multipart
aiofiles
# motor # Removed since it's a MongoDB dependency
# pydantic==2.5.3 # Remove strict version constraint
# pydantic-settings # Remove since it's dependent on pydantic
fastapi
websockets # Add websockets for real-time updates
supabase # Add Supabase client for Python
